MainSolution/requirements.txt
Flask
flask-cors
pywhatkit
python-dotenv
pydantic
requests
python-docx
